+++ lib/mjml_eex.ex
@@ -1,7 +1,31 @@
 defmodule MjmlEEx do
   @moduledoc """
   Documentation for `MjmlEEx` template module. This moule contains the macro
-  that is used to create an MJML EEx template.
+  that is used to create an MJML EEx template. The macro can be configured to
+  render the MJML template in a few different ways, so be sure to read the
+  option documentation.
+
+  ## Macro Options
+
+  - `:mjml_template`- A binary that specifies the name of the `.mjml.eex` template that the module will compile. The
+    directory path is relative to the template module. If this option is not provided, the MjmlEEx will look for a
+    file that has the same name as the module but with the `.mjml.ex` extension as opposed to `.ex`.
+
+  - `:mode`- This option defines when the MJML template is actually compiled. The possible values are `:runtime` and
+    `:compile`. When this option is set to `:compile`, the MJML template is compiled into email compatible HTML at
+    compile time. It is suggested that this mode is only used if the template is relatively simple and there are only
+    assigns being used as text or attributes on html elements (as opposed to attributes on MJML elements). The reason
+    for that being that these assigns may be discarded as part of the MJML compilation phase. On the plus side, you
+    do get a performance bump here since the HTML for the email is already generated. When this is set to `:runtime`,
+    the MJML template is compiled at runtime and all the template assigns are applied prior to the MJML compilation
+    phase. These means that there is a performance hit since you are compiling the MJML template every time, but the
+    template can use more complex EEx constructs like `for`, `case` and `cond`. The default configuration is `:runtime`.
+
+  - `:layout` - This option defines what layout the template should be injected into prior to rendering the template.
+    This is useful if you want to have reusable email templates in order to keep your email code DRY and reusable.
+    Your template will then be injected into the layout where the layout defines `<%= inner_content %>`.
+
+  ## Example Usage
 
   You can use this module like so:
 
@@ -37,18 +61,14 @@ defmodule MjmlEEx do
   alias MjmlEEx.Utils
 
   defmacro __using__(opts) do
-    mjml_template =
-      case Keyword.fetch(opts, :mjml_template) do
-        {:ok, mjml_template} ->
-          %Macro.Env{file: calling_module_file} = __CALLER__
+    # Get some data about the calling module
+    %Macro.Env{file: calling_module_file} = __CALLER__
+    module_directory = Path.dirname(calling_module_file)
+    file_minus_extension = Path.basename(calling_module_file, ".ex")
+    mjml_template_file = Keyword.get(opts, :mjml_template, "#{file_minus_extension}.mjml.eex")
 
-          calling_module_file
-          |> Path.dirname()
-          |> Path.join(mjml_template)
-
-        :error ->
-          raise "The :mjml_template option is required."
-      end
+    # The absolute path of the mjml template
+    mjml_template = Path.join(module_directory, mjml_template_file)
 
     unless File.exists?(mjml_template) do
       raise "The provided :mjml_template does not exist at #{inspect(mjml_template)}."
@@ -65,13 +85,10 @@ defmodule MjmlEEx do
     raw_mjml_template =
       case layout_module do
         :none ->
-          get_raw_template(mjml_template, __CALLER__)
+          get_raw_template(mjml_template, compilation_mode, __CALLER__)
 
         module when is_atom(module) ->
-          get_raw_template_with_layout(mjml_template, layout_module, __CALLER__)
-
-        invalid_layout ->
-          raise "#{inspect(invalid_layout)} is an invalid layout option"
+          get_raw_template_with_layout(mjml_template, layout_module, compilation_mode, __CALLER__)
       end
 
     generate_functions(compilation_mode, raw_mjml_template, mjml_template, layout_module)
@@ -159,18 +176,18 @@ defmodule MjmlEEx do
     raise "#{inspect(invalid_mode)} is an invalid :mode. Possible values are :runtime or :compile"
   end
 
-  defp get_raw_template(template_path, caller) do
+  defp get_raw_template(template_path, mode, caller) do
     {mjml_document, _} =
       template_path
       |> File.read!()
       |> Utils.escape_eex_expressions()
-      |> EEx.compile_string(engine: MjmlEEx.Engines.Mjml, line: 1, trim: true, caller: caller)
+      |> EEx.compile_string(engine: MjmlEEx.Engines.Mjml, line: 1, trim: true, caller: caller, mode: mode)
       |> Code.eval_quoted()
 
     Utils.decode_eex_expressions(mjml_document)
   end
 
-  defp get_raw_template_with_layout(template_path, layout_module, caller) do
+  defp get_raw_template_with_layout(template_path, layout_module, mode, caller) do
     template_file_contents = File.read!(template_path)
     pre_inner_content = layout_module.pre_inner_content()
     post_inner_content = layout_module.post_inner_content()
@@ -179,7 +196,7 @@ defmodule MjmlEEx do
       [pre_inner_content, template_file_contents, post_inner_content]
       |> Enum.join()
       |> Utils.escape_eex_expressions()
-      |> EEx.compile_string(engine: MjmlEEx.Engines.Mjml, line: 1, trim: true, caller: caller)
+      |> EEx.compile_string(engine: MjmlEEx.Engines.Mjml, line: 1, trim: true, caller: caller, mode: mode)
       |> Code.eval_quoted()
 
     Utils.decode_eex_expressions(mjml_document)