+++ test/mjml_eex_test.exs
@@ -19,6 +19,18 @@ defmodule MjmlEExTest do
       mode: :compile
   end
 
+  defmodule DynamicComponentTemplate do
+    use MjmlEEx,
+      mjml_template: "test_templates/dynamic_component_template.mjml.eex",
+      mode: :runtime
+  end
+
+  defmodule InvalidDynamicComponentTemplate do
+    use MjmlEEx,
+      mjml_template: "test_templates/invalid_dynamic_component_template.mjml.eex",
+      mode: :runtime
+  end
+
   defmodule FunctionTemplate do
     use MjmlEEx,
       mjml_template: "test_templates/function_template.mjml.eex",
@@ -94,11 +106,31 @@ defmodule MjmlEExTest do
         end
       end
     end
+
+    test "should raise an error if the MJML template compile mode is invalid" do
+      assert_raise RuntimeError, ~r/:yolo is an invalid :mode. Possible values are :runtime or :compile/, fn ->
+        defmodule InvalidCompileModeOption do
+          use MjmlEEx,
+            mjml_template: "test_templates/invalid_template.mjml.eex",
+            mode: :yolo
+        end
+      end
+    end
+
+    test "should raise an error if the layout option is invalid" do
+      assert_raise ArgumentError, ~r/could not load module InvalidModule due to reason/, fn ->
+        defmodule InvalidLayoutOption do
+          use MjmlEEx,
+            mjml_template: "test_templates/invalid_template.mjml.eex",
+            layout: InvalidModule
+        end
+      end
+    end
   end
 
   describe "The use macro" do
-    test "should fail to compile since a required option is not present" do
-      assert_raise RuntimeError, ~r/The :mjml_template option is required./, fn ->
+    test "should fail to compile since a valid mjml template can not be found" do
+      assert_raise RuntimeError, ~r/The provided :mjml_template does not exist at/, fn ->
         defmodule NoTemplateOption do
           use MjmlEEx
         end
@@ -123,15 +155,67 @@ defmodule MjmlEExTest do
     end
   end
 
+  describe "DynamicComponentTemplate.render/1" do
+    test "should render the document with the appropriate assigns" do
+      rendered_template = DynamicComponentTemplate.render(some_data: 1..5)
+
+      assert rendered_template =~ "Some data - 1"
+      assert rendered_template =~ "Some data - 2"
+      assert rendered_template =~ "Some data - 3"
+      assert rendered_template =~ "Some data - 4"
+      assert rendered_template =~ "Some data - 5"
+    end
+  end
+
+  describe "CompileTimeDynamicComponentTemplate.render/1" do
+    test "should raise an error if a dynamic component is rendered at compile time" do
+      assert_raise RuntimeError,
+                   ~r/render_dynamic_component can only be used with runtime generated templates. Switch your template to `mode: :runtime`/,
+                   fn ->
+                     defmodule CompileTimeDynamicComponentTemplate do
+                       use MjmlEEx,
+                         mjml_template: "test_templates/dynamic_component_template.mjml.eex",
+                         mode: :compile
+                     end
+                   end
+    end
+  end
+
+  describe "InvalidDynamicComponentTemplate.render/1" do
+    test "should raise an error as dynamic components cannot render other dynamic components" do
+      assert_raise RuntimeError,
+                   ~r/Cannot call `render_dynamic_component` inside of another dynamically rendered component/,
+                   fn ->
+                     InvalidDynamicComponentTemplate.render(some_data: 1..5)
+                   end
+    end
+  end
+
+  describe "BadExpressionDynamicComponentTemplate" do
+    test "should fail to compile since the render_dynamic_component call is not in an = expression" do
+      assert_raise RuntimeError,
+                   ~r/render_dynamic_component can only be invoked inside of an <%= ... %> expression/,
+                   fn ->
+                     defmodule BadExpressionDynamicComponentTemplate do
+                       use MjmlEEx,
+                         mjml_template: "test_templates/bad_expression_dynamic_component_template.mjml.eex",
+                         mode: :runtime
+                     end
+                   end
+    end
+  end
+
   describe "InvalidComponentTemplate" do
-    test "should fail to compile since the render_component call is not in an = expression" do
-      assert_raise RuntimeError, ~r/render_component can only be invoked inside of an <%= ... %> expression/, fn ->
-        defmodule InvalidTemplateOption do
-          use MjmlEEx,
-            mjml_template: "test_templates/invalid_component_template.mjml.eex",
-            mode: :compile
-        end
-      end
+    test "should fail to compile since the render_static_component call is not in an = expression" do
+      assert_raise RuntimeError,
+                   ~r/render_static_component can only be invoked inside of an <%= ... %> expression/,
+                   fn ->
+                     defmodule InvalidTemplateOption do
+                       use MjmlEEx,
+                         mjml_template: "test_templates/invalid_component_template.mjml.eex",
+                         mode: :compile
+                     end
+                   end
     end
   end