+++ test/test_components/invalid_dynamic_component.ex
@@ -1,0 +1,16 @@
+defmodule MjmlEEx.TestComponents.InvalidDynamicComponent do
+  @moduledoc """
+  This module defines the MJML component for the shared head block.
+  """
+
+  use MjmlEEx.Component
+
+  @impl true
+  def render(data: data) do
+    """
+    <p>
+      <%= render_dynamic_component MjmlEEx.TestComponents.DynamicComponent %>
+    </p>
+    """
+  end
+end